ingestor/requirements.txt
redis
confluent-kafka
mysql-connector-python
